pyproject.toml
[tool.poetry]
name = "test_projects_from_hr"
version = "0.1.0"
description = ""
authors = ["dopefresh <dopefresh4000@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "2.2.10"
environs = "^9.3.2"
djangorestframework = "^3.12.4"
loguru = "^0.5.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
